Created scraping folder.
Scraping is a part of Axelbib which is using the python library
arcas: to scrape spesific apis and post the result to Axelbib.
Script.py is the script which does that and it also slips put
a report file and a failed validation file.

Minnor corrections and alterations were done.

Finally Analysis and Results contains various notebooks used for the analysis
of the database of Axelbib.